+++ src/main/java/com/jnape/palatable/lambda/lens/functions/Over.java
@@ -4,7 +4,7 @@ import com.jnape.palatable.lambda.functions.Fn1;
 import com.jnape.palatable.lambda.functions.Fn2;
 import com.jnape.palatable.lambda.functions.Fn3;
 import com.jnape.palatable.lambda.functor.builtin.Identity;
-import com.jnape.palatable.lambda.lens.Lens;
+import com.jnape.palatable.lambda.lens.LensLike;
 
 import java.util.function.Function;
 
@@ -23,7 +23,7 @@ import java.util.function.Function;
  * @see Set
  * @see View
  */
-public final class Over<S, T, A, B> implements Fn3<Lens<S, T, A, B>, Function<? super A, ? extends B>, S, T> {
+public final class Over<S, T, A, B> implements Fn3<LensLike<S, T, A, B, ?>, Function<? super A, ? extends B>, S, T> {
 
     private static final Over INSTANCE = new Over();
 
@@ -31,10 +31,8 @@ public final class Over<S, T, A, B> implements Fn3<Lens<S, T, A, B>, Function<? 
     }
 
     @Override
-    public T apply(Lens<S, T, A, B> lens, Function<? super A, ? extends B> fn, S s) {
-        return lens.<Identity, Identity<T>, Identity<B>>fix()
-                .apply(fn.andThen((Function<B, Identity<B>>) Identity::new), s)
-                .runIdentity();
+    public T apply(LensLike<S, T, A, B, ?> lens, Function<? super A, ? extends B> fn, S s) {
+        return lens.<Identity, Identity<T>, Identity<B>>apply(fn.andThen((Function<B, Identity<B>>) Identity::new), s).runIdentity();
     }
 
     @SuppressWarnings("unchecked")
@@ -42,17 +40,15 @@ public final class Over<S, T, A, B> implements Fn3<Lens<S, T, A, B>, Function<? 
         return (Over<S, T, A, B>) INSTANCE;
     }
 
-    public static <S, T, A, B> Fn2<Function<? super A, ? extends B>, S, T> over(
-            Lens<S, T, A, B> lens) {
+    public static <S, T, A, B> Fn2<Function<? super A, ? extends B>, S, T> over(LensLike<S, T, A, B, ?> lens) {
         return Over.<S, T, A, B>over().apply(lens);
     }
 
-    public static <S, T, A, B> Fn1<S, T> over(Lens<S, T, A, B> lens,
-                                              Function<? super A, ? extends B> fn) {
+    public static <S, T, A, B> Fn1<S, T> over(LensLike<S, T, A, B, ?> lens, Function<? super A, ? extends B> fn) {
         return over(lens).apply(fn);
     }
 
-    public static <S, T, A, B> T over(Lens<S, T, A, B> lens, Function<? super A, ? extends B> fn, S s) {
+    public static <S, T, A, B> T over(LensLike<S, T, A, B, ?> lens, Function<? super A, ? extends B> fn, S s) {
         return over(lens, fn).apply(s);
     }
 }